StarlarkTypeRepr::Canonical
Summary:
If `A` and `B` represent the same starlark type, `A::Canonical == B::Canonical`. For example `String::Canonical == <&str>::Canonical`.

This is useful with `ValueOfUnchecked::cast` in D59380757.

This diff unfortunately makes starlark API harder to use.

It would be better if we had [`associated_type_defaults`](https://github.com/rust-lang/rust/issues/29661), but that won't happen soon.

Reviewed By: JakobDegen

Differential Revision: D59380699

fbshipit-source-id: 8b658203c2d46cd03ac722170ba3858d19eee001