derive: Remove overly aggressive spans
Summary:
So, it is kind of correct that these produce slightly improved diagnostics if the proc macro has a bug. However, if it doesn't, these are actively harmful for IDEs.

Broadly speaking, the way that r-a works with proc macros is that if the macro generated code references a symbol, and that symbol has some span on it, then the associated span in the source code is treated as referencing that symbol. As a result, over-aggressively applying spans to generated code means much of the stuff referenced by the generated code shows up on hover and go-to-def where it shouldn't.

Reviewed By: stepancheg

Differential Revision: D60276898

fbshipit-source-id: 5a6a66aed575a9e6dd5728b10db48e7489e7caf5