Remove nop `#[derive(StarlarkDocs)]`
Summary:
Let me first give a bit of context: I want to get rid of `#[derive(StarlarkDocs)]` for generating documentation. Instead, I want to use only the list of globals, generating docs for a type if an appropriate type symbol appears in the globals.

This diff just removes invocations of the derive macro that didn't do anything - specifically, the derive macro only produces global docs if the associated type has any method defined on it, and in all the examples here that was not the case.

https://fburl.com/code/jrf6aakj

Reviewed By: stepancheg

Differential Revision: D60686908

fbshipit-source-id: f86556e0378d0c93306c644e50237080f5e9dc46