docs: Nest `DocItem`s in `DocModule`s
Summary:
It makes sense for modules to be able to contain other modules and objects.

Unfortunately, the status quo of the way that types are used in this code is quite poor, which means that it's hard to write this diff in a way that is obviously correct. For most of the existing codepaths, we handle the new branches by throwing out modules and collapsing objects to properties, which will generally have the effect of preserving behavior when we start to use this

Reviewed By: stepancheg

Differential Revision: D60686916

fbshipit-source-id: d7e0021ca4d2a0bd1d359d68c33fb2637f7dfba9