docs: Copy of structs for generating json
Summary:
`buck2 docs starlark` defaults to outputting json. It does this by just putting `Serialize` on all the types in the starlark crate and then printing that out. Unfortunately, that makes those types basically impossible to change because everything about them is now a part of the stable API of buck2.

So let's avoid that by making a copy of the types so that we can keep the json output fixed while changing the underlying representation.

This is kind of a lot of boilerplate but I don't really see how to get around that.

Reviewed By: stepancheg

Differential Revision: D60686888

fbshipit-source-id: f55b22d7a55362719cc518f3ddbea02e5b8ff2a3