Resolve doc_lazy_continuation warnings in documentation
Summary:
This lint caught a legit mistake in the documentation.

```lang=text,counterexample
error: doc list item missing indentation
  --> superconsole/src/lib.rs:20:5
   |
20 | //! Components live in the scratch area.
   |     ^
   |
   = help: if this is supposed to be its own paragraph, add a blank line
   = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html...
   = note: `-D clippy::doc-lazy-continuation` implied by `-D clippy::all`
   = help: to override `-D clippy::all` add `#[allow(clippy::doc_lazy_continuation)]`
help: indent this line
   |
20 | //!   Components live in the scratch area.
   |     ++
```

Without a blank line between the last bullet and next line, according to Markdown that sentence becomes part of the last bullet. Based on the content, it definitely does not seem like that was intended by the author of this documentation.

 {F1795256035}

Reviewed By: shayne-fletcher, JakobDegen

Differential Revision: D60933067

fbshipit-source-id: 7aba872b716c6351f000d77f44ac0e4b5f292765