Bump rustyline dep to 14.0.0 (#123)
Summary:
This allows one to have `nix` 0.28 instead of 0.26 in one's dep graph, so that platform support of `rustyline` becomes better. No behavioral change intended for starlark-rust because the dependence on rustyline is minimal.

Note though that nix's MSRV is raised to 1.65 starting from 0.27.0, and so does starlark-rust.

Pull Request resolved: https://github.com/facebook/starlark-rust/pull/123

Reviewed By: zertosh

Differential Revision: D61429152

Pulled By: ndmitchell

fbshipit-source-id: 149bcee4c91d5def7942933a2f673e5da66a05d4