Add a WASM/JS example (#129)
Summary:
Based on the code in https://github.com/facebook/starlark-rust/issues/109#issuecomment-2225024605, with a few minor changes to make it work with stable and work for types that aren't string. Plus a slightly nicer HTML page where you can edit the input.

![image](https://github.com/user-attachments/assets/cc5a1f42-8e25-4284-9d9c-9ccffdeab318)

Credit for most of the code goes to aschleck.

Pull Request resolved: https://github.com/facebook/starlark-rust/pull/129

Reviewed By: stepancheg

Differential Revision: D61774902

Pulled By: ndmitchell

fbshipit-source-id: 4babe21de199f194be3405cd7e22b543a3e9169d