NativeCallableParams
Summary:
This diff adds structs

```
pub struct NativeCallableParam {
    pub name: &'static str,
    pub ty: Ty,
    pub required: bool,
}

pub struct NativeCallableParamSpec {
    pub pos_only: Vec<NativeCallableParam>,
    pub pos_or_named: Vec<NativeCallableParam>,
    pub args: Option<NativeCallableParam>,
    pub named_only: Vec<NativeCallableParam>,
    pub kwargs: Option<NativeCallableParam>,
}
```

populated by proc macros. So in `Ty::from_native_callable_components` we don't use `ParametersSpec`.

Context: `ParametersSpec` is for runtime. We should not be using it in typechecking (this diff), and in documentation generation (to be done).

Next step would be to
- add default value to `NativeCallableParam`
- use it to generate documentation
- remove `NativeCallableComponents.signature` (which is after this diff is only used for documentation)

Reviewed By: JakobDegen

Differential Revision: D63014287

fbshipit-source-id: 4cfec47663a8fbb6460ff5f54d147972c60012b2