Add buck2_error Context for Starlark Backtraces
Summary:
Will be needed in `starlark:Error->buck2_error::Error` conversion in the next diff. Starlark errors will have a backtrace which contain useful information that buck2_error cannot easily capture right now.

This is sort of a hack because if there are multiple stack traces it will just combine them together. Combine stack traces properly is difficult and this should be good enough for now

Reviewed By: christolliday

Differential Revision: D64123339

fbshipit-source-id: 4d6d66ad8d38301c65ff1522c15c81c5eeaaee67