starlark::Error <-> buck2_error conversion
Summary:
Add implementations for buck2_error<->starlark. Code will be used in the next diff.

Note that there's `from_starlark` and `from_starlark_impl` because we need a way for caller to choose to not print stacktrace and/or be categorized (Though the categorized should eventually be removed once we categorize everything properly, but we will do this for now)

Reviewed By: JakobDegen

Differential Revision: D63487582

fbshipit-source-id: cdfa3ea9b8cd42a1d1428ff8420be8a96ee83a76