docs: Don't collapse docs for global items
Summary:
Currently, `DocItem::Type` and `DocItem::Module`, when they appear in globals, just get transformed into `DocMember::Property` before reaching `Globals::documentation`. Change that.

This is a slight regression in documentation quality for now, because `DocModule`s are now ignored completely instead of getting rendered as the very marginally more useful `name: namespace(..)`. We'll have to live with that for a few diffs, it'll be fixed properly in this stack

Reviewed By: stepancheg

Differential Revision: D61778037

fbshipit-source-id: 2ca720c170974ffa9176feeaf2fc3f1e92d86829