docs: Flip constructor handling
Summary:
In starlark, we have a number of globals that act as both functions and types - `str`, `int`, but also all the providers and more. Currently, the docs just render these things as functions but have a little indicator of a `.type` attribute on them. This diff flips things around: Document these globals primarily as types instead, and add the documentation for the constructor function to the type docs.

Example for `str`:

{F1834529832}

There are some cases in which this is less useful. For example, `int` doesn't have any inherent methods and so the type page ends up just documenting the constructor:

{F1834533828}

We could try and avoid that, but honestly I think it's fine. Thoughts welcome.

I'll also note that builtin providers don't have their constructor function documented. That's because they currently do a bunch of their own doc magic which doesn't support this yet. However, I think that's actually ok and we should keep it that way. The constructor function doesn't really add any additional information on top of what is already visible from the attributes:

{F1834539663}

This is undoing Neil's change from D45230206 , but I think the context has changed a little bit to make that ok. We're now consistent about documenting types in their own page, and nowhere else

Reviewed By: cjhopman

Differential Revision: D64858308

fbshipit-source-id: 9f30e171413a8f041e90cfdf63e4ba00c2ca24fd