Fix Go-to behavior for prelude & native symbols on Windows
Summary:
Context: VS Code's Starlark extension allows you to navigate to definitions of prelude and native symbols. However, this go-to behavior seems to have been broken for a while (or perhaps didn't work at all) on Windows machines.

After investigating, I noticed that the issue is due to misformatted pathing. Windows expects delimiters of "/", not "\" (as was previously given). So, for example, instead of sending a request with uri `starlark:/native%5Cgetattr.bzl`, we send a request with uri `starlark:/native/getattr.bzl`, which is able to properly be computed and navigated to on Windows.

Thankfully, turns out to be a quick one liner to enable the go-to behavior for Windows!

Reviewed By: ndmitchell, LoxFalcon, JakobDegen

Differential Revision: D68602359

fbshipit-source-id: 1687a948b440766f58a4e1087c124800790e10f5