Rename functions to better reflect their purpose
Summary:
This is prep work for D69315461.

Today, profilers either produce flame data or csv data. D69315461 will allow profilers to produce both, which means the functionality in the `.gen()` method needs to be split into per-output methods, so that both can be called on the profilers that support both.

Reviewed By: cjhopman

Differential Revision: D69368277

fbshipit-source-id: 29c4e72b06e5fe2147a226a8634b753aa385c9a9