StrongHash: allow passing in a regular Hasher
Summary:
StrongHash mixes two different concerns:

- Arbitrary output size (StrongHasher).
- Not using precomputed weak hashes (StrongHash).

However, you can want the regular hash output size (aka 64 bits) while not wanting to
use precomputed weak hashes, but right now that's not allowed. This is typically
useful if you are passing your thing to a library that wants a regular 64 bit hash.

This diff fixes that by allowing you to pass a regular `Hasher` into
`fn strong_hash`, and also updates the docs to express the notion that it is OK
to want to use StrongHash even if your hasher isn't a StrongHasher.

Note: ideally you wouldn't want to let someone use a `StrongHasher` and then
pass it into `hash` instead of `strong_hash`, but that was already possible
before, and it doesn't seem like it's been a problem.

Reviewed By: cjhopman

Differential Revision: D73095294

fbshipit-source-id: 80e14308c8547f82b84e25fd20c573f5a98e8cd6