improve statement profiler accuracy
Summary:
Prior, we kept a `BcStatementLocation` that tracked which BC instruction was the first associated with a statement. This is used for a few things, including breakpoint evaluation, statement profiling, etc.

For statement profiling, this would lead to bytecode like this for a statement misreporting all of the time it takes to `Call Bar` as time spent in "whatever the last first-statement instruction was executed during the Call Foo->Return".
```
    A: Add
    B: Call Foo
    C: Call Bar
```

Previously, only `A` here was marked in BcStatementLocations. Now we mark `A` and `C` as belonging to the statement, too. This causes extra `before_stmt` calls that can then do the additional lookups/accounting in the profiler to credit the correct statement with the time to execute the instruction at `C`.

`BcStatementLocations` still differentiates between the "first" instruction for a statement, vs all the others (the `continued` ones), so that components like the debug adaptor can still "break on next statement", rather than "next statement-annotated instruction".

Reviewed By: cjhopman

Differential Revision: D73038348

fbshipit-source-id: 6804c86ffb2a082b8848b99850dd75e0d6ba1afa