module: Hand write virtualization for `NativeAttribute`
Summary:
So Rust's type system has a, uhh, limitation:

```
fn foo<T>(){}

fn call_static<U: Fn() + 'static>(u: U) {
    u();
}

pub fn check<'x>() {
    // Doesn't work: lifetime may not live long enough
    // call_static(foo::<&'x ()>);
    call_static(foo::<&'x ()> as fn());
}
```

Function ZSTs, when instantiated with some non-static generics, are not `'static`, while function pointers are.

I'm looking to add generics support to `#[starlark_module]`, so think of the `foo` function as being the implementation for the function/attribute that the user wrote.

This causes problems. `NativeAttribute` values need to be `'static`-ish, but they also need access to generic parameters that might be very non-static as a result of being `StarlarkValue`s.

Working around the limitation above would require me to convert to a function pointer - which I could do, but that would introduce a double virtual call. So to avoid introducing the double virtual call, we explicit split out the `&dyn NativeAttr` into a data value and function pointer. That way, we can use the function pointer that we must explicitly create above to do the single virtual function invocation

Hopefully that made sense

Reviewed By: Nero5023

Differential Revision: D73726014

fbshipit-source-id: 4ee8f71cda1e0cc7bd10e60d1f270b766825efa5