Make Evaluator covariant over 'a
Summary:
'a represents the lifetimes of references that the Evaluator captures.
The Evaluator is currently invariant over this lifetime and it can lead to some
hard to understand and hard to resolve lifetime issues. There's no real reason
for the invariance, so this fixes it and adds assertions that it continues to hold.

Reviewed By: JakobDegen

Differential Revision: D75649198

fbshipit-source-id: eda0eb3cb29c683f0c8f90ea0ff5e5c4acd5e381