Updated approach for with_starlark_eval_provider
Summary:
Before this, starlark evaluation is wrapped in something like:

```
with_starlark_eval_provider(
    <args>,
    closure: impl for<'y> FnOnce(&mut StarlarkEvaluatorProvider<'y, 'a>, D) -> buck2_error::Result<R>)
```

And the StarlarkEvaluatorProvider had a function to make a starlark Evaluator.

With this, you could only hold the StarlarkEvaluatorProvider through sync code,
and we were doing unexpected things like returning the created Evaluator
out of the with_starlark_eval_provider to get around that. Also that structure
limited what interesting things we could move into the common code.

This diff changes the structure so that now we have StarlarkEvaluatorProvider
as a struct that users create that provides a `with_evaluator(<args>, impl
FnOnce(&mut Evaluator<'v, 'a2, 'e2>, bool) -> buck2_error::Result<R>)` and
the evaluator provider (or a ReentrantStarlarkEvaluator) is the only thing that
owns the Evaluator.

This will allow moving a lot more things into common code so that they can be
handled consistently across all invocations. Specifically for this diff stack,
that will mean supporting a consistent mechanism to enable profiling on any
evaluation.

Reviewed By: JakobDegen

Differential Revision: D64281588

fbshipit-source-id: e925e29f7ff78d3320fdd191c591e3d8e801d949